requirements_base.txt
Babel==2.3.4
Django==1.8.8
IPy==0.83
Pygments==2.1.3
Twisted==16.2.0
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
asgi-redis==0.14.1
asgiref==0.14.0
attrs==16.2.0
autobahn==0.16.0
backports.shutil-get-terminal-size==1.0.0
boto==2.42.0
channels==0.17.2
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
daphne==0.14.3
decorator==4.0.10
dj-database-url==0.4.1
django-countries==4.0
django-easymoney==0.7.1
django-extensions==1.6.7
django-floppyforms==1.7.0
django-idmap==0.4.1
django-inspect-model==0.7
django-vanilla-views==1.0.4
djangorestframework==3.4.6
funcsigs==1.0.2
honcho==0.7.1
huey==1.2.0
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
mock==2.0.0
msgpack-python==0.4.8
ordered-set==2.0.1
otree-core==0.9.8
otree-save-the-change==1.1.3
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.5
psutil==4.2.0
ptyprocess==0.5.1
py==1.4.31
pytest==2.9.2
pytest-django==3.0.0
pytz==2016.6.1
raven==5.25.0
redis==2.10.5
requests==2.11.1
schema==0.6.2
selenium==2.53.2
simplegeneric==0.8.1
six==1.10.0
tqdm==4.7.4
traitlets==4.2.2
txaio==2.5.1
unicodecsv==0.14.1
wcwidth==0.1.7
wheel==0.29.0
whitenoise==3.2.1
ws4py==0.3.5
wsgiref==0.1.2
xmltodict==0.10.1
zope.interface==4.2.0
